requirements.txt
streamlit-mic-recorder==0.0.4
InstructorEmbedding==1.0.1
ctransformers==0.2.27
transformers==4.38.1
langchain==0.1.9
langchain-community==0.0.22
streamlit==1.31.1
pypdfium2==4.27.0
librosa==0.10.1
pyyaml==6.0.1

